Move `Cow::clear` methods near `ToOwned` impls
Applies to `Cow<[T]>` and `Cow<str>`. `Cow<OsStr>` is already like this.

This allows us to use intra-doc links with ease since the relevant items
are locally in scope.
