+++ library/alloc/src/borrow.rs
@@ -328,71 +328,6 @@ impl<B: ?Sized + ToOwned> Cow<'_, B> {
     }
 }
 
-impl<T> Cow<'_, [T]>
-where
-    T: Clone,
-{
-    /// Truncates the [`Vec`] or [`&[T]`](prim@slice).
-    ///
-    /// See [`Vec::clear`].
-    ///
-    /// # Examples
-    ///
-    /// Basic usage:
-    ///
-    /// ```
-    /// #![feature(cow_clear)]
-    /// use std::borrow::Cow;
-    ///
-    /// let mut owned: Cow<'_, [i32]> = Cow::Owned(vec![1, 2, 3]);
-    /// owned.clear();
-    /// assert!(owned.is_empty());
-    ///
-    /// let mut borrowed = Cow::Borrowed([1, 2, 3].as_slice());
-    /// borrowed.clear();
-    /// assert!(borrowed.is_empty());
-    /// ```
-    #[inline]
-    #[unstable(feature = "cow_clear", issue = "none")]
-    pub fn clear(&mut self) {
-        match self {
-            Self::Owned(vec) => vec.clear(),
-            Self::Borrowed(slice) => *slice = &[],
-        }
-    }
-}
-
-impl Cow<'_, str> {
-    /// Truncates the [`String`] or [`&str`](prim@str).
-    ///
-    /// See [`String::clear`].
-    ///
-    /// # Examples
-    ///
-    /// Basic usage:
-    ///
-    /// ```
-    /// #![feature(cow_clear)]
-    /// use std::borrow::Cow;
-    ///
-    /// let mut owned: Cow<'_, str> = Cow::Owned(String::from("foo"));
-    /// owned.clear();
-    /// assert!(owned.is_empty());
-    ///
-    /// let mut borrowed = Cow::Borrowed("bar");
-    /// borrowed.clear();
-    /// assert!(borrowed.is_empty());
-    /// ```
-    #[inline]
-    #[unstable(feature = "cow_clear", issue = "none")]
-    pub fn clear(&mut self) {
-        match self {
-            Self::Owned(string) => string.clear(),
-            Self::Borrowed(str) => *str = "",
-        }
-    }
-}
-
 #[stable(feature = "rust1", since = "1.0.0")]
 impl<B: ?Sized + ToOwned> Deref for Cow<'_, B>
 where
